Refactor client factory and streamline deprecated modules.
Reorganized the client factory structure, deprecating the old implementation and enhancing code modularity. Removed unused code, standardized file naming, and addressed separation of concerns. Added deprecation warnings for obsolete modules, migrating towards more robust implementations.
