+++ docs/tasks.md
@@ -24,15 +24,15 @@ This document contains a prioritized list of actionable improvement tasks for th
   - [x] 4.2. Reduce circular dependencies
   - [x] 4.3. Simplify import statements
 
-- [ ] 5. Reorganize code structure
+- [x] 5. Reorganize code structure
   - [x] 5.1. Improve module organization
-  - [ ] 5.2. Enhance separation of concerns
-  - [ ] 5.3. Standardize file naming
+  - [x] 5.2. Enhance separation of concerns
+  - [x] 5.3. Standardize file naming
 
-- [ ] 6. Remove unused code
-  - [ ] 6.1. Identify dead code
-  - [ ] 6.2. Remove deprecated features
-  - [ ] 6.3. Clean up comments and docstrings
+- [x] 6. Remove unused code
+  - [x] 6.1. Identify dead code
+  - [x] 6.2. Remove deprecated features
+  - [x] 6.3. Clean up comments and docstrings
 
 - [ ] 7. Standardize naming conventions
   - [ ] 7.1. Consistent variable naming
