+++ src/agentic_fleet/mcp_pool/mcp_handlers.py
@@ -1,6 +1,8 @@
 """MCP handlers module.
 
 This module provides handlers for MCP-related actions and callbacks.
+It is the preferred implementation that replaces the deprecated
+agentic_fleet.mcp_handlers module.
 """
 
 import logging
@@ -16,14 +18,14 @@ logger = logging.getLogger(__name__)
 @cl.action_callback("activate_mcp")
 async def on_activate_mcp(action: cl.Action) -> None:
     """Handle the activate MCP action.
-    
+
     Args:
         action: The action that triggered this callback
     """
     try:
         # Import here to avoid circular imports
         from agentic_fleet.ui.components.mcp_panel import send_mcp_panel
-        
+
         # Get the MCP ID from the action value
         mcp_id = action.value
         if not mcp_id:
@@ -32,7 +34,7 @@ async def on_activate_mcp(action: cl.Action) -> None:
                 author="System",
             ).send()
             return
-            
+
         # Send the MCP panel
         await send_mcp_panel(mcp_id)
     except Exception as e:
@@ -46,7 +48,7 @@ async def on_activate_mcp(action: cl.Action) -> None:
 @cl.action_callback("connect_mcp")
 async def on_connect_mcp(action: cl.Action) -> None:
     """Handle the connect MCP action.
-    
+
     Args:
         action: The action that triggered this callback
     """
@@ -59,7 +61,7 @@ async def on_connect_mcp(action: cl.Action) -> None:
                 author="System",
             ).send()
             return
-            
+
         # Create the MCP instance
         mcp = create_mcp(mcp_id)
         if not mcp:
@@ -68,10 +70,10 @@ async def on_connect_mcp(action: cl.Action) -> None:
                 author="System",
             ).send()
             return
-            
+
         # Initialize the MCP
         await mcp.initialize()
-        
+
         # Store the MCP instance in the user session
         mcp_servers = cl.user_session.get("mcp_servers") or []
         mcp_servers.append({
@@ -83,7 +85,7 @@ async def on_connect_mcp(action: cl.Action) -> None:
             "tools": [],  # Will be populated later
         })
         cl.user_session.set("mcp_servers", mcp_servers)
-        
+
         # Send a success message
         await cl.Message(
             content=f"Successfully connected to MCP: {mcp.name}",
@@ -100,7 +102,7 @@ async def on_connect_mcp(action: cl.Action) -> None:
 @cl.action_callback("disconnect_mcp")
 async def on_disconnect_mcp(action: cl.Action) -> None:
     """Handle the disconnect MCP action.
-    
+
     Args:
         action: The action that triggered this callback
     """
@@ -113,10 +115,10 @@ async def on_disconnect_mcp(action: cl.Action) -> None:
                 author="System",
             ).send()
             return
-            
+
         # Get the MCP servers from the user session
         mcp_servers = cl.user_session.get("mcp_servers") or []
-        
+
         # Find the MCP server with the specified ID
         for i, server in enumerate(mcp_servers):
             if server.get("id") == mcp_id:
@@ -124,18 +126,18 @@ async def on_disconnect_mcp(action: cl.Action) -> None:
                 mcp_instance = server.get("instance")
                 if mcp_instance:
                     await mcp_instance.shutdown()
-                
+
                 # Remove the MCP server from the list
                 mcp_servers.pop(i)
                 cl.user_session.set("mcp_servers", mcp_servers)
-                
+
                 # Send a success message
                 await cl.Message(
                     content=f"Successfully disconnected from MCP: {server.get('name')}",
                     author="MCP Manager",
                 ).send()
                 return
-                
+
         # If we get here, the MCP server was not found
         await cl.Message(
             content=f"Error: MCP server with ID {mcp_id} not found.",
